refactor: separate core funcs
